port: transition Marble to Qualcomm 5.15.149 and validate boot prelinks
Base the source snapshot on kernel/msm 603278d412c33acb62666bf8e09ccfc0b8732165, retaining Marble/Cape board data and reviewed initial adapters. Preserve the unified 5.10 reference at cd3f531d2230a22812f6c3f0ffe0148928d70d03.

Record native LLVM 18 compilation and 26 module-sized prelinks, plus a controlled DT warning comparison. No full vmlinux, final modules, package or device validation is claimed. Generated artifacts and downloaded tooling are excluded.

+++ arch/m68k/include/asm/bitops.h
@@ -314,12 +314,12 @@ static inline int bfchg_mem_test_and_change_bit(int nr,
 #include <asm-generic/bitops/ffz.h>
 #else
 
-static inline unsigned long find_first_zero_bit(const unsigned long *vaddr,
-						unsigned long size)
+static inline int find_first_zero_bit(const unsigned long *vaddr,
+				      unsigned size)
 {
 	const unsigned long *p = vaddr;
-	unsigned long res = 32;
-	unsigned long words;
+	int res = 32;
+	unsigned int words;
 	unsigned long num;
 
 	if (!size)
@@ -340,9 +340,8 @@ out:
 }
 #define find_first_zero_bit find_first_zero_bit
 
-static inline unsigned long find_next_zero_bit(const unsigned long *vaddr,
-					       unsigned long size,
-					       unsigned long offset)
+static inline int find_next_zero_bit(const unsigned long *vaddr, int size,
+				     int offset)
 {
 	const unsigned long *p = vaddr + (offset >> 5);
 	int bit = offset & 31UL, res;
@@ -371,12 +370,11 @@ static inline unsigned long find_next_zero_bit(const unsigned long *vaddr,
 }
 #define find_next_zero_bit find_next_zero_bit
 
-static inline unsigned long find_first_bit(const unsigned long *vaddr,
-					   unsigned long size)
+static inline int find_first_bit(const unsigned long *vaddr, unsigned size)
 {
 	const unsigned long *p = vaddr;
-	unsigned long res = 32;
-	unsigned long words;
+	int res = 32;
+	unsigned int words;
 	unsigned long num;
 
 	if (!size)
@@ -397,9 +395,8 @@ out:
 }
 #define find_first_bit find_first_bit
 
-static inline unsigned long find_next_bit(const unsigned long *vaddr,
-					  unsigned long size,
-					  unsigned long offset)
+static inline int find_next_bit(const unsigned long *vaddr, int size,
+				int offset)
 {
 	const unsigned long *p = vaddr + (offset >> 5);
 	int bit = offset & 31UL, res;
@@ -442,8 +439,6 @@ static inline unsigned long ffz(unsigned long word)
 }
 
 #endif
-
-#include <asm-generic/bitops/find.h>
 
 #ifdef __KERNEL__
 
@@ -528,10 +523,12 @@ static inline int __fls(int x)
 #define __clear_bit_unlock	clear_bit_unlock
 
 #include <asm-generic/bitops/ext2-atomic.h>
-#include <asm-generic/bitops/le.h>
 #include <asm-generic/bitops/fls64.h>
 #include <asm-generic/bitops/sched.h>
 #include <asm-generic/bitops/hweight.h>
+#include <asm-generic/bitops/le.h>
 #endif /* __KERNEL__ */
+
+#include <asm-generic/bitops/find.h>
 
 #endif /* _M68K_BITOPS_H */
